car/schema.py
import graphene
from graphene_django import DjangoObjectType

from car.models import Category, Ingredient, Car


class CarType(DjangoObjectType):
    class Meta:
        model = Car
        fields = ("id", "name", "color", "brand")

class CategoryType(DjangoObjectType):
    class Meta:
        model = Category
        fields = ("id", "name", "ingredients")

class IngredientType(DjangoObjectType):
    class Meta:
        model = Ingredient
        fields = ("id", "name", "notes", "category")

class Query(graphene.ObjectType):
    all_car = graphene.List(CarType)
    all_ingredients = graphene.List(IngredientType)
    category_by_name = graphene.Field(CategoryType, name=graphene.String(required=True))


    def resolve_all_car(root, info):
        return Car.objects.all()

    def resolve_all_ingredients(root, info):
        # We can easily optimize query count in the resolve method
        return Ingredient.objects.all()

    def resolve_category_by_name(root, info, name):
        try:
            return Category.objects.get(name=name)
        except Category.DoesNotExist:
            return None

schema = graphene.Schema(query=Query)